+++ processor/src/test/fixtures/expected/com/example/prop/React4j_ObservableProp.java
@@ -82,7 +82,8 @@ abstract class React4j_ObservableProp extends ObservableProp {
   abstract ObservableValue<?> getValueObservableValue();
 
   @Action(
-      verifyRequired = false
+      verifyRequired = false,
+      reportParameters = false
   )
   boolean $$react4j$$_shouldComponentUpdate(@Nullable final JsPropertyMap<Object> nextInputs) {
     assert null != nextInputs;